+++ src/pools/selection-strategies/selection-strategies-types.ts
@@ -46,6 +46,12 @@ export interface WorkerChoiceStrategyOptions {
    */
   medRunTime?: boolean
   /**
+   * Use tasks median wait time instead of average runtime.
+   *
+   * @defaultValue false
+   */
+  medWaitTime?: boolean
+  /**
    * Worker weights to use for weighted round robin worker selection strategy.
    * Weight is the tasks maximum average or median runtime in milliseconds.
    *
@@ -72,6 +78,18 @@ export interface RequiredStatistics {
    * Require tasks median runtime.
    */
   medRunTime: boolean
+  /**
+   * Require tasks wait time.
+   */
+  waitTime: boolean
+  /**
+   * Require tasks average wait time.
+   */
+  avgWaitTime: boolean
+  /**
+   * Require tasks median wait time.
+   */
+  medWaitTime: boolean
 }
 
 /**